glew

+++ TurboTribble/Engine/Source/glmath.cpp
@@ -24,7 +24,7 @@ vec2 mix(const vec2 &u, const vec2 &v, float a)
 
 vec2 normalize(const vec2 &u)
 {
-	return u / sqrt(u.x * u.x + u.y * u.y);
+	return u / sqrt((double)u.x * (double)u.x + (double)u.y * (double)u.y);
 }
 
 vec2 reflect(const vec2 &i, const vec2 &n)
